requirements.txt
feedparser
google-genai
python-dotenv
markdown
groq
arxiv